Move primitive parsing methods to a separate module
We want to be able to test the parsing logic separate from an actual HTML page. This step extracts all parsing methods that operate on primitives to a separate module so that we can test that module in isolation
