+++ src/setup.js
@@ -7,6 +7,7 @@ const Stats = require('./stats');
 const WorkerStats = require('./worker-stats');
 const Scenario = require("./scenario");
 const LineChart = require("./charts");
+const {parseInput} = require("./parsing");
 
 function createScenarioContainer(scenario) {
     const template = document.querySelector('#scenario-template');
@@ -16,22 +17,6 @@ function createScenarioContainer(scenario) {
     clone.querySelector('.scenario-title').textContent = scenario.title;
     return clone
 }
-
-function parseWorkload(input) {
-  return input
-    .trim()
-    .split(',')
-    .map(pair => pair
-      .trim()
-      .split(":")
-    )
-    .reduce((work, pair) => {
-      work[pair[0].trim()] = parseInt(pair[1].trim());
-      return work;
-    }, {})
-}
-
-const split = value => value.trim().split(",").map(item => item.trim());
 
 function parse(form) {
   const field = fieldName => form.querySelector(`[name="${fieldName}"]`).value;
@@ -43,27 +28,6 @@ function parse(form) {
       wipLimit: field('wip-limit'),
       random: form.querySelector('[name="random"]').checked
   });
-}
-
-function parseInput(rawInput) {
-    const title = rawInput.title;
-    const workers = split(rawInput.workers);
-    const work = parseWorkload(rawInput.workload);
-    const wipLimit = rawInput.wipLimit;
-    const speed = (workers.length > 2) ? 20 : 1;
-    const numberOfStories = (workers.length > 2) ? 200 : 50;
-    let input = {
-        title,
-        workers,
-        stories: {
-            amount: numberOfStories,
-            work
-        },
-        wipLimit,
-        speed
-    };
-    if (rawInput.random) input.distribution = average
-    return input;
 }
 
 function parseScenario(event) {
@@ -90,7 +54,6 @@ document.addEventListener('DOMContentLoaded', () => {
 const wipLimiter = LimitBoardWip();
 
 
-const {average} = require("./generator");
 let currentChart = undefined;
 
 function run(scenario) {
